+++ src/app/home/FollowMissionList.tsx
@@ -1,5 +1,6 @@
 import Link from 'next/link';
 import { useFollowMissions } from '@/apis/mission';
+import { MissionListSkeleton } from '@/app/home/home.styles';
 import MissionBadge from '@/app/home/MissionBadge';
 import MissionEmptyList from '@/app/home/MissionEmptyList';
 import { TwoLineListItem } from '@/components/ListItem';
@@ -7,8 +8,6 @@ import { MISSION_CATEGORY_LABEL } from '@/constants/mission';
 import { ROUTER } from '@/constants/router';
 import { css } from '@styled-system/css';
 import { flex } from '@styled-system/patterns';
-
-import { Skeleton } from './MissionList';
 
 interface FollowMissionListProps {
   followId: number;
@@ -45,7 +44,7 @@ const listCss = flex({
 function MissionFollowListInner({ followId }: { followId: number }) {
   const { data, isLoading } = useFollowMissions(followId ?? 0);
   if (isLoading) {
-    return <Skeleton />;
+    return <MissionListSkeleton />;
   }
 
   if (!data) {
